style: Add selected state (outer circle) to the color picking palette

+++ src/tool/base/colorparabar.h
@@ -41,7 +41,7 @@ protected:
     virtual void paintEvent(QPaintEvent *event) override;
 
 private:
-    double m_scal;
+    int m_scal;
     Qt::Orientations m_orien;
     QGridLayout* m_layout;
     QMap<QString, QString> m_labMap;
