+++ src/tool/selectbar.h
@@ -60,7 +60,7 @@ signals:
     void sigFinish();
 
 private:
-    double m_scal;
+    int m_scal;
     Qt::Orientations m_orien;
     QBoxLayout* m_layout;
